Rename 'NotificationTokenHelper' into 'NotificationCenterToken'.
